Use 'sweep' for filenames, too.
